Major Overhaulv2
Was able to work the problems out that I was having that constituted using the test file. Redeveloped so that test was no longer needed. Each GUI component is its own class again. Implemented slidepane in center feature so that Actions tabs could have a good size while still giving the majority of the pane to the content console.