Security Changes and Started Actions
Realized that there was a potential security risk with the Users text file. Redid the 365 logins to create users.txt upon start of sequence and delete at end of sequence. Also started brainstorming actions to include. If list of actions becomes too large, will change pane to slidePane as needed.